Move IP checks to conf.py.
